Refactor cacheables
