+++ app/controllers/AmendAreYouSureController.scala
@@ -16,13 +16,12 @@
 
 package controllers
 
-import cacheables.ObligationCacheable
+import cacheables.{AmendReturnPreviousReturn, AmendSelectedPeriodKey, ObligationCacheable}
 import connectors.CacheConnector
-import controllers.ViewReturnSummaryController.{AmendReturnPreviousReturn, AmendSelectedPeriodKey}
 import controllers.actions._
 import controllers.helpers.TaxReturnHelper
 import forms.AmendAreYouSureFormProvider
-import models.{Mode, UserAnswers}
+import models.Mode
 import navigation.Navigator
 import pages.AmendAreYouSurePage
 import play.api.data.Form
